Fix tls link write cb (#59)
* tls_link: write_cb should be called after actual write
* prevent writes to closed websockets
* prevent writes to closed websockets
* set closed if websocket src couldnt conenct
* use correct link in write_cb

+++ include/uv_mbed/tls_engine.h
@@ -81,7 +81,7 @@ typedef struct {
       * @param out_bytes
       * @param maxout
       */
-    int (*write)(void *engine, const char *data, size_t data_len, char *out, size_t *out_bytes, size_t maxout);
+    ssize_t (*write)(void *engine, const char *data, size_t data_len, char *out, size_t *out_bytes, size_t maxout);
 
     /**
      * process bytes received from TLS peer. Application data is placed in out buffer.
